+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # v2.0.0
 
+## Major changes
+* Merged `filtering` with some types from `logging` as `config`.
+* Moved `prettylogger::logging::Logger` to `prettylogger::Logger`.
+
 ## Minor changes
-* Added a new function in `colors` module that colors text using ANSII escape
-codes.
+* Added a new function in `colors` module that can be used to color text using
+ANSII escape codes.
+* Added an option to toggle logging to `stdout`.
 
 ## Patches
 * Changed `prettylogger::setters` module accessibility to private.